+++ api/src/prescription/campaign-participation/application/api/campaign-participations-api.js
@@ -10,15 +10,3 @@ import { usecases } from '../../domain/usecases/index.js';
 export const hasCampaignParticipations = async ({ userId }) => {
   return usecases.hasCampaignParticipations({ userId });
 };
-
-/**
- * @function
- * @name deleteAllForOrganizationLearner
- *
- * @param {number} organizationLearnerId
- * @param {number} userId
- * @returns {Promise<void>}
- */
-export const deleteAllForOrganizationLearner = async function ({ organizationLearnerId, userId }) {
-  return usecases.deleteCampaignParticipationsForOrganizationLearner({ organizationLearnerId, userId });
-};
