+++ api/src/prescription/learner-management/domain/usecases/index.js
@@ -1,11 +1,13 @@
 import { dirname, join } from 'node:path';
 import { fileURLToPath } from 'node:url';
 
+import * as badgeAcquisitionRepository from '../../../../evaluation/infrastructure/repositories/badge-acquisition-repository.js';
 import * as userRepository from '../../../../identity-access-management/infrastructure/repositories/user.repository.js';
 import * as organizationFeatureApi from '../../../../organizational-entities/application/api/organization-features-api.js';
 import * as obfuscationService from '../../../../shared/domain/services/obfuscation-service.js';
 import * as placementProfileService from '../../../../shared/domain/services/placement-profile-service.js';
 import * as userReconciliationService from '../../../../shared/domain/services/user-reconciliation-service.js';
+import { featureToggles } from '../../../../shared/infrastructure/feature-toggles/index.js';
 import { logErrorWithCorrelationIds } from '../../../../shared/infrastructure/monitoring-tools.js';
 import * as libOrganizationLearnerRepository from '../../../../shared/infrastructure/repositories/organization-learner-repository.js';
 import * as organizationRepository from '../../../../shared/infrastructure/repositories/organization-repository.js';
@@ -14,6 +16,7 @@ import { importNamedExportsFromDirectory } from '../../../../shared/infrastructu
 import { logger } from '../../../../shared/infrastructure/utils/logger.js';
 import * as membershipRepository from '../../../../team/infrastructure/repositories/membership.repository.js';
 import * as campaignRepository from '../../../campaign/infrastructure/repositories/campaign-repository.js';
+import * as campaignParticipationRepositoryfromBC from '../../../campaign-participation/infrastructure/repositories/campaign-participation-repository.js';
 import * as registrationOrganizationLearnerRepository from '../../../organization-learner/infrastructure/repositories/registration-organization-learner-repository.js';
 import { repositories } from '../../infrastructure/repositories/index.js';
 import { importCommonOrganizationLearnersJobRepository } from '../../infrastructure/repositories/jobs/import-common-organization-learners-job-repository.js';
@@ -29,7 +32,6 @@ import * as organizationLearnerRepository from '../../infrastructure/repositorie
 import * as studentRepository from '../../infrastructure/repositories/student-repository.js';
 import * as supOrganizationLearnerRepository from '../../infrastructure/repositories/sup-organization-learner-repository.js';
 import { importStorage } from '../../infrastructure/storage/import-storage.js';
-
 /**
  * @typedef {import ('../../infrastructure/repositories/organization-feature-repository.js')} CampaignParticipationRepository
  * @typedef {import ('../../../campaign/infrastructure/repositories/campaign-repository.js')} CampaignRepository
@@ -69,6 +71,7 @@ const dependencies = {
   logErrorWithCorrelationIds,
   logger,
   membershipRepository,
+  campaignParticipationRepositoryfromBC,
   obfuscationService,
   organizationFeatureApi,
   organizationFeatureRepository: repositories.organizationFeatureRepository,
@@ -81,10 +84,12 @@ const dependencies = {
   studentRepository,
   supOrganizationLearnerRepository,
   userReconciliationService,
+  badgeAcquisitionRepository,
   userRepository,
   validateCommonOrganizationImportFileJobRepository,
   validateCsvOrganizationImportFileJobRepository,
   validateOrganizationImportFileJobRepository,
+  featureToggles,
 };
 
 const path = dirname(fileURLToPath(import.meta.url));
